+++ src/gfx/window.hpp
@@ -11,8 +11,8 @@ class Window{
         Camera* camera;
 
         char* name;
-        int SCR_WIDTH;
-        int SCR_HEIGHT;
+        float SCR_WIDTH;
+        float SCR_HEIGHT;
 
         bool firstMouse;
         float yaw;     //left / right direction. Starting at -90 to face straight?
@@ -21,12 +21,11 @@ class Window{
         float lastX;
         float lastY;
 
-        Window(const int &SCR_WIDTH, const int &SCR_HEIGHT, Camera* camera);
+        Window(const float &SCR_WIDTH, const float &SCR_HEIGHT, Camera* camera);
         bool init();
 
         void setCallbacks();
         void processInput(Camera* camera, const float& deltaTime);
-
     private:
         static void framebuffer_size_callback(GLFWwindow *window, int width, int height);
         static void scroll_callback(GLFWwindow *window, double xoffset, double yoffset);
